+++ src/Application/SmartHome.Application.Shared/Events/Devices/GreenhouseController/Telemetry/GreenhouseControllerTelemetryEvent.cs
@@ -6,9 +6,9 @@ namespace SmartHome.Application.Shared.Events.Devices.GreenhouseController.Telem
     {
         public DateTime MeasurementStartTime { get; set; }
         public DateTime MeasurementEndTime { get; set; }
-        public float Temperature { get; set; }
-        public float Pressure { get; set; }
-        public float Humidity { get; set; }
+        public double Temperature { get; set; }
+        public double Pressure { get; set; }
+        public double Humidity { get; set; }
         public int LightLevel { get; set; }
         public int SoilMoisture { get; set; }
         public bool IsDoorOpen { get; set; }
