+++ src/Domain/Entities/Devices/Greenhouse/GreenhouseIrrigationHistory.cs
@@ -6,10 +6,10 @@ namespace SmartHome.Domain.Entities.Devices.Greenhouse
     public class GreenhouseIrrigationHistory: IMeasurement
     {
         public long Id { get; set; }        
-        public float TotalWaterVolume { get; set;}
-        public float AverageWaterFlow { get; set;}
-        public float MinWaterFlow { get; set;}
-        public float MaxWaterFlow { get; set;}
+        public double TotalWaterVolume { get; set;}
+        public double AverageWaterFlow { get; set;}
+        public double MinWaterFlow { get; set;}
+        public double MaxWaterFlow { get; set;}
         public DateTime MeasurementStartTime { get; set; }
         public DateTime MeasurementEndTime { get; set; }
     }
